+++ ChangeLog.md
@@ -4,6 +4,7 @@
 * Fix handling of UIDs in DicomCFindRequest
 * Fix comparison of private DicomTags
 * Add shortcut constructor for private DicomTags
+* Handle null DicomDateRange in DicomDataset.Add()
 
 #### v1.0.14 (09/05/2012)
 * Fix bug in DicomDatasetReaderObserver handling zero length Specific Character Set elements